requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.10.5
numpy==2.2.2
Pillow==10.4.0
opencv-python==4.10.0.84
insightface==0.7.3
onnxruntime==1.22.1
starlette==0.37.2
cryptography==42.0.8
